test: Control which files are converted to PDF
In order to do better testing of the PDF to HTML conversion, we'll need some files that have been converted to PDF via different means as well as matching HTML files to test against. This means we won't be able to convert every HTML file to PDF as this would override any PDF files we want to generate with different tools.